pyproject.toml
[tool.poetry]
name = "forestbot"
version = "0.1.0"
description = ""
authors = ["Guillaume Soto", "Tony Lopes" ,"Connor Haines"]
readme = "README.md"
packages = [{include = "src/api/*.py"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.4"
Flask = "^3.0.3"
openai = "^1.17.1"
langchain = {extras = ["docarray"], version = "^0.1.16"}
langchain-openai = "^0.1.3"
python-dotenv = "^1.0.1"
panel = "^1.4.1"
param = "^2.1.0"
ipykernel = "^6.29.4"
pypdf = "^4.2.0"
streamlit = "^1.33.0"
langchain-community = "^0.0.33"
chromadb = "^0.4.24"
streamlit-chat = "^0.1.1"
icecream = "^2.1.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"
ruff = "^0.3.7"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"